Ok so Im having a massive mental block rn but let me explain
Ive been steadily improving the player controller over the week and have made a lot of progress, the player can move, face wherever the mouse is aiming, attack more precisely thanks to the raycasts  detecting if anything attackable is in range. Recently I started working on its health logic and attacking but to truly work on it I had to begin work on what will be our games first enemy, I call them The shattered and while I wont go in depth rn all I wanted to do with it for now is give it the ability to hurt/kill the player and be killed by the player but im having some trouble because of the sensory overload of info and data I have to mess with so yeah pls help lol. If you need help to understand what we need to do call me but I need a bit of mental stimulation cuz I dont know how you psychos can sit here programming without talking to anybody for hours lol.