We can now apply different damage values to enemy's to damage them (Still needs more work)
So ive been working on our first enemy and at first it was a struggle but ive so far gotten rid of some mayor issues when it came to actually damaging them. So it started with me wondering how we would get the enemy to know its being hit by a weapon with a different damage value to the one it got hit with first, so I began work on a method that detects this change, it uses enumerators (enum) from another script in which all these damage types will be located to identify what value and calculation will be used (note that as of right now there is only 2 and the value is being applied from the enemy script by using switch cases to switch in between which value is used along with the calculation to subtract health). The goal will be to overhaul this later to grab variables from the weapons of the game to set its damage values, effects, multipliers, etc. into more switch cases using other enumerators and determine how it affects the enemy.

As of right now there are still bugs like for example the multiple raycasts to detect enemies also individually apply the damage so if you have 5 raycast colliding with the player it will apply the damage from each one causing more damage than it should, theres also a frame issue in which the damage will apply very quickly over and over again as the frames cycle, so if you click and it cycles through 10 frames in that second that you clicked it will then apply that damage 10 times leading to the same problem from before, there may be more but until further testing I cant fully say.

You can apply the punch damage (10 dmg) if you left click and slash damage (20 dmg) with right click, these were only used as testing values and are prone to change.